Better subdivision and naming
