+++ collectives/models/__init__.py
@@ -9,9 +9,8 @@ from collectives.models.activity_type import ActivityType, leaders_without_activ
 from collectives.models.auth import ConfirmationToken, ConfirmationTokenType
 from collectives.models.configuration import ConfigurationItem, ConfigurationTypeEnum
 from collectives.models.configuration import Configuration
-from collectives.models.event import Event, EventStatus, photos
+from collectives.models.event import Event, EventStatus, photos, EventType
 from collectives.models.event_tag import EventTag
-from collectives.models.event_type import EventType
 from collectives.models.equipment import EquipmentModel, Equipment, EquipmentType
 from collectives.models.equipment import image_equipment_type, EquipmentStatus
 from collectives.models.payment import PaymentItem, ItemPrice, PaymentStatus
